continue correcting modbus server

+++ decode_config/generate_global_buffer_multi_buffer.py
@@ -106,9 +106,10 @@ typedef struct ModbusBlocks{
     int LastTagIndex;
 }ModbusBlocks;
 
+// modbus tags have not status and date. so we need to delete them from driverTag structure. 
 typedef struct Modbus_master_driverTag{
-    double TagValue;
-    double OldTagValue;
+    long double TagValue;
+    long double OldTagValue;
     uint32_t TagStatus;
     uint32_t OldTagStatus;
     IEC_DT TagValueDT;
@@ -126,8 +127,8 @@ typedef struct Modbus_master_driverTag{
 } Modbus_master_driverTag;
 
 typedef struct Modbus_slave_driverTag{
-    double TagValue;
-    double OldTagValue;
+    long double TagValue;
+    long double OldTagValue;
     uint32_t TagStatus;
     uint32_t OldTagStatus;
     IEC_DT TagValueDT;
@@ -162,8 +163,8 @@ typedef struct Database_driver_options{
 }Database_driver_options;
 
 typedef struct Database_driverTag{
-    double TagValue;
-    double OldTagValue;
+    long double TagValue;
+    long double OldTagValue;
     uint32_t TagStatus;
     uint32_t OldTagStatus;
     IEC_DT TagValueDT;
@@ -192,8 +193,8 @@ typedef struct LIO_driver_options{
 }LIO_driver_options;
 
 typedef struct LIO_driverTag{
-    double TagValue;
-    double OldTagValue;
+    long double TagValue;
+    long double OldTagValue;
     uint32_t TagStatus;
     uint32_t OldTagStatus;
     IEC_DT TagValueDT;
@@ -283,8 +284,8 @@ typedef struct DNP_slave_driver_options{
 }DNP_slave_driver_options;
 
 typedef struct DNP_master_driverTag{
-    double TagValue;
-    double OldTagValue;
+    long double TagValue;
+    long double OldTagValue;
     uint32_t TagStatus;
     uint32_t OldTagStatus;
     IEC_DT TagValueDT;
@@ -302,8 +303,8 @@ typedef struct DNP_master_driverTag{
 } DNP_master_driverTag;
 
 typedef struct DNP_slave_driverTag{
-    double TagValue;
-    double OldTagValue;
+    long double TagValue;
+    long double OldTagValue;
     uint32_t TagStatus;
     uint32_t OldTagStatus;
     IEC_DT TagValueDT;
